% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chf_1F1_gsl}
\alias{chf_1F1_gsl}
\title{Kummer's (confluent hypergeometric) function}
\usage{
chf_1F1_gsl(x, a, b, log_v = FALSE)
}
\arguments{
\item{x}{numeric value or vector}

\item{a, b}{numeric parameters of the Kummer function}

\item{log_v}{logical; if TRUE, the log of the value is returned}
}
\description{
Kummer's (confluent hypergeometric) function of the first kind
for numeric (non-complex) values and input parameters
}
