% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNA_basic.R
\name{gRNA_basic}
\alias{gRNA_basic}
\title{Gillespie Algorithm to simulate from basic model of mRNA generation and degradation}
\usage{
gRNA_basic(n, r, mu)
}
\arguments{
\item{n}{number of simulations}

\item{r}{mRNA generation rate}

\item{mu}{mRNA degradation rate}
}
\description{
Gillespie Algorithm to simulate from basic model of mRNA generation and degradation
}
\examples{
gRNA_basic(1000,  r = 10, mu = 1)


}
