% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb.R
\name{Inverse Gaussian}
\alias{Inverse Gaussian}
\alias{rInvGaus}
\title{Inverse Gaussian Distribution}
\usage{
rInvGaus(n, mu, lambda)
}
\arguments{
\item{n}{Number of observations}

\item{mu, lambda}{Non-negative parameters of the inverse Gaussian distribution (mean and shape)}
}
\description{
random generation function for the inverse Gaussian distribution:
Mu and lambda are the parameters of this distribution.
}
\examples{
 RV <- rInvGaus(n = 100, mu = 10, lambda = 2)
}
\keyword{Gaussian}
\keyword{Inverse}
\keyword{distribution}
