% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kummer}
\alias{kummer}
\alias{kummer_gsl}
\title{Kummer function for real values}
\usage{
kummer_gsl(x, a, b, lnchf = 0L)
}
\arguments{
\item{x}{vector of parameters}

\item{a}{floating point parameter}

\item{b}{floating point parameter}

\item{lnchf}{0 or 1; if 1, the log of the value is returned}
}
\description{
Kummer function for real values
}

