% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypergeometricFunctions.R, R/RcppExports.R
\name{KummerM}
\alias{KummerM}
\alias{kummerM}
\alias{kummer_gsl}
\title{The kummer function}
\usage{
kummerM(x, a, b, lnchf = 0, ip = 0)

kummer_gsl(x, a, b, lnchf = 0L)
}
\arguments{
\item{x}{complex function argument}

\item{a}{complex index a}

\item{b}{complex index b}

\item{lnchf}{an integer indicating if the output is logarithmic or not}

\item{ip}{an integer indicating the number of terms to compute in
the kummer series}
}
\description{
A function implemented by Diethelm Wuertz
Calculate the Confluent Hypergeometric Function of the First Kind
The kummer_gsl uses the GNU Scientific Library implementation.
}

