% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodNB.R
\name{likelihood-nb-mpb}
\alias{likelihood-nb-mpb}
\alias{Loglik_nb}
\alias{Loglik_mpb}
\alias{Loglik_nb_zero}
\alias{Loglik_mpb_zero}
\title{Likelihood functions for negative binomial and mixed poisson-beta}
\usage{
Loglik_nb(data, par.nb)

Loglik_mpb(data, par.mpb)

Loglik_nb_zero(data, par.nb)

Loglik_mpb_zero(data, par.mpb)
}
\arguments{
\item{data}{The dataset for which the likelihood is computed}

\item{par.nb}{Parameters for the negative binomial distribution}

\item{par.mpb}{Parameters for the mixed-poisson-beta distribution}
}
\description{
The likelihood functions are required for optimisation purposes.
There are functions to compute the likelihood of the negative
binomial and the mixed poisson-beta distributions with and
without a zero inflation.
}
\keyword{binomial}
\keyword{likelihood}
\keyword{mixed}
\keyword{negative}
\keyword{poisson-beta}
