% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpb2.R
\name{mpb}
\alias{dmpb}
\alias{mpb}
\alias{pmpb}
\alias{qmpb}
\alias{rmpb}
\title{Mixed Poisson-Beta Distribution}
\usage{
dmpb(x, alpha, beta, c, log = FALSE)

pmpb(q, alpha, beta, c, lower.tail = TRUE, log.p = FALSE)

qmpb(p, alpha, beta, c, lower.tail = TRUE, log.p = FALSE)

rmpb(n, alpha, beta, c)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{alpha, beta}{non-negative parameters of the beta distribution (shape1 and shape2)}

\item{c}{numeric scaling parameter of the beta distribution
(i.e. the standard beta(alpha,beta) distribution is scaled not on (0, 1), but on (0, c))}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
Density, distribution function, quantile function and random generation
for the mixed Poisson-beta distribution. This represents a Poisson
distribution whose parameter itself follows a beta distribution. This
beta distribution uses the parameters alpha and beta. Additionally this
beta distribution is scaled on (0, c) in contrast to the usual scaling on (0,1).
}
\examples{
 X <- dmpb(x=0:200, alpha=5, beta=3, c=20)
 plot(0:200, X, type='l')
 Y <- dmpb(0:10, seq(10.0,11.0,by=0.1), seq(30.0,31.0,by=0.1), seq(10.2,11.2,by=0.1))
 Y <- pmpb(q= 0 :200, alpha=5, beta= 3, c=20)
 plot(0:200, Y, type="l")
 Z <- qmpb(p= seq(0,1, by= 0.01), alpha=5, beta= 3, c=20)
 plot(seq(0,1, by= 0.01),Z, type="l")
 RV <- rmpb(n = 1000, alpha=5, beta= 3, c=20)
 plot(0 : 200, X, type="l")
 lines(density(RV), col="red")
 R2 <- rmpb(11, seq(10.0,11.0,by=0.1), seq(30.0,31.0,by=0.1), seq(10.2,11.2,by=0.1))
}
\keyword{distribution}
\keyword{mixed}
\keyword{poisson-beta}

