% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mpb2.R
\name{dmpb}
\alias{dmpb}
\alias{mpb2}
\alias{rmpb}
\title{Mixed Poisson-Beta Distribution}
\usage{
dmpb(x, alpha, beta, c)

rmpb(n, alpha, beta, c)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{alpha}{first non-negative parameter of the beta distribution (shape1)}

\item{beta}{second non-negative parameter of the beta distribution b(shape2)}

\item{c}{thirs parameter of the beta distribution (i.e. the standard
beta(alpha,beta) distribution is scaled not on (0, 1), but on (0, c.value))}

\item{n}{number of observations}

\item{p}{vector of probabilities}
}
\description{
Density, distribution function, quantile function and random
generation for the mixed poisson-beta distribution. This is basically a
poisson distribution where the parameter itself is again beta distributed
with parameters alpha and beta and additional scaled on (0, c).
In difference to the usual scaling on (0,1)
}
\examples{
 X <- dmpb(x=0:200, alpha=5, beta=3, c=20)
 plot(0:200, X, type='l')
 RV <- rMPB(n = 1000, alpha=5, beta= 3, c=20)

 plot(0 : 200, X, type="l")
 lines(density(RV), col="red")
}
\keyword{distribution}
\keyword{mixed}
\keyword{poisson-beta}

