% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_est_fns.R
\name{par-est-fns}
\alias{par-est-fns}
\alias{estimate_mpb_optim_init}
\alias{estimate_mpb_optim_init_restarts}
\alias{get_0inf_parameter}
\alias{get_fitted_params}
\title{Functions to compute parameter estimates}
\usage{
estimate_mpb_optim_init(x, iter = 200)

estimate_mpb_optim_init_restarts(x, n = 10)

get_0inf_parameter(x)

get_fitted_params(x, type, optim_contol = list())
}
\arguments{
\item{x}{Gene expression data as an array}

\item{iter}{number of bootstrap replicates to estimate initial
parameters for the mpb}

\item{type}{keyword for the distribution the data is to be fitted
against. Possible values are ("pois", "zip", "nb", "zinb",
"mpb", "zimpb", "pois2", "nb2", "mpb2")}

\item{optim_contol}{list of options to override default settings in
the optim function; only valid types: mpb, zimpb and mpb2.}
}
\description{
Functions to compute parameter estimates
}
\keyword{estimation}
\keyword{parameter}
