% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_est_fns.R
\name{fit_params}
\alias{fit_params}
\title{Functions to fit distributions}
\usage{
fit_params(x, type, optim_contol = list())
}
\arguments{
\item{x}{Vector of count data}

\item{type}{keyword for the distribution the data is to be fitted
against. Possible values are ("pois", "nb", "pb", "pois2", "nb2",
 "pb2", "zipois", "zinb", "zipb", "zipois2", "zinb2", "zipb2")}

\item{optim_contol}{list of options to override default settings in
the optim function; only valid types: "pb", "pb2", "zipb", "zipb2".
For more detials, please refer to the 'control' parameter in the
standard 'optim' function in package 'stats'.}
}
\description{
Functions to fit distributions
}
\keyword{estimation}
\keyword{parameter}
