% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmRNA}
\alias{gmRNA}
\title{Gillespie Algorithm to simulate from basic kinetic model of gene activation and mRNA transcription}
\usage{
gmRNA(n, lambda, gamma, r, mu)
}
\arguments{
\item{n}{number of simulations}

\item{lambda}{polymerase binding rate / DNA activation rate}

\item{gamma}{polymerase unbinding rate / DNA deactivation rate}

\item{r}{transcription rate}

\item{mu}{mRNA degradation rate}
}
\description{
Gillespie Algorithm to simulate from basic kinetic model of gene activation and mRNA transcription
}

