% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gillespie.R
\name{gmRNA}
\alias{gmRNA}
\alias{gmRNA_basic}
\alias{gmRNA_switch}
\alias{gmRNA_burst}
\title{Gillespie algorithm for mRNA generating processes}
\usage{
gmRNA_basic(n, r.on, r.degr)

gmRNA_switch(n, r.act, r.deact, r.on, r.degr)

gmRNA_burst(n, r.burst, s.burst, r.degr)
}
\arguments{
\item{n}{Number of observations}

\item{r.on}{Transcription rate during gene activation (Switching model)}

\item{r.degr}{mRNA degradation rate (all models)}

\item{r.act}{DNA activation rate (Switching Model)}

\item{r.deact}{DNA deactivation rate (Switching Model)}

\item{r.burst}{Bursty transcription rate (Bursting model)}

\item{s.burst}{Mean burst size (Bursting Model)}
}
\description{
Gillespie algorithms allow synthetic data simulation via three different
underlying mRNA generating processes: the basic process consists of a
simple death-birth model of mRNA transcription and degradation; the
switching process considers additionally gene activation and deactivation,
with mRNA transcription only happening in active gene states; the
bursting process, transcribes mRNA in bursts with geometrically distributed burst sizes.
}
