% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{likelihood-nb-mpb}
\alias{likelihood-nb-mpb}
\alias{nLoglik_pois}
\alias{nLoglik_nb}
\alias{nLoglik_mpb}
\alias{nLoglik_pois_zero}
\alias{nLoglik_nb_zero}
\alias{nLoglik_mpb_zero}
\alias{nLoglik_pois_two}
\alias{nLoglik_nb_two}
\alias{nLoglik_mpb_two}
\title{Likelihood functions for negative binomial and mixed poisson-beta}
\usage{
nLoglik_pois(data, par.pois)

nLoglik_nb(data, par.nb)

nLoglik_mpb(data, par.mpb)

nLoglik_pois_zero(data, par.pois.zero)

nLoglik_nb_zero(data, par.nb.zero)

nLoglik_mpb_zero(data, par.mpb.zero)

nLoglik_pois_two(data, par.pois2)

nLoglik_nb_two(data, par.nb2)

nLoglik_mpb_two(data, par.mpb2)
}
\arguments{
\item{data}{The dataset for which the likelihood is computed}

\item{par.pois}{Parameters for the Poisson distribution}

\item{par.nb}{Parameters for the negative binomial distribution}

\item{par.mpb}{Parameters for the mixed-poisson-beta distribution}

\item{par.pois.zero, par.nb.zero, par.mpb.zero}{Parameters for the
respective zero-inflated distributions. The first in the array
is always the inflation parameter for all cases.}

\item{par.pois2, par.nb2, par.mpb2}{Parameters for the respective
two population distributions}
}
\description{
The likelihood functions are required for optimisation purposes.
There are functions to compute the likelihood of the negative
binomial and the mixed poisson-beta distributions with and
without a zero inflation.
}
\details{
Functions Loglik_pois, Loglik_nb, Loglik_mpb compute the negative
log-likelihood of poisson, negative binomial and the mixed poisson-beta
distributions against the data.
Functions Loglik_pois_zero, Loglik_nb_zero, Loglik_mpb_zero compute the
zero-inflated log-likelihood values.
}
\keyword{binomial}
\keyword{likelihood}
\keyword{mixed}
\keyword{negative}
\keyword{poisson-beta}
