% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{nlogL}
\alias{nlogL}
\alias{nlogL_pois}
\alias{nlogL_nb}
\alias{nlogL_pb}
\alias{nlogL_pois2}
\alias{nlogL_nb2}
\alias{nlogL_pb2}
\alias{nlogL_zipois}
\alias{nlogL_zinb}
\alias{nlogL_zipb}
\alias{nlogL_zipois2}
\title{Negative log Likelihood functions for Poisson, negative binomial and
Poisson-beta distributions}
\usage{
nlogL_pois(data, par.pois)

nlogL_nb(data, par.nb)

nlogL_pb(data, par.pb)

nlogL_pois2(data, par.pois2)

nlogL_nb2(data, par.nb2)

nlogL_pb2(data, par.pb2)

nlogL_zipois(data, par.zipois)

nlogL_zinb(data, par.zinb)

nlogL_zipb(data, par.zipb)

nlogL_zipois2(data, par.zipois2)
}
\arguments{
\item{data}{Vector containing the discrete observations}

\item{par.pois}{Scalar containing the lambda parameter
of the Poisson distribution}

\item{par.nb}{Vector of length 2, containing the size and the mu
parameter of the negative binomial distribution}

\item{par.pb}{Vector of length 3, containing the alpha, beta
and c parameter of the Poisson-beta distribution}

\item{par.pois2, par.nb2, par.pb2}{Vector containing the parameters
of the two mixing distributions. First entry represents the
fraction of the first distribution, followed by all parameters
of the first, then all of the second distribution.}

\item{par.zipois, par.zinb, par.zipb}{Vector containing the respective
zero-inflated distribution parameters. The additional first
entry is the inflation parameter for all cases.}

\item{par.zipois2}{Parameters for the zero-inflated 2 population
model.}
}
\description{
The negative log Likelihood functions for Poisson, negative binomial
and Poisson-beta distributions. Mixing two distributions of the same
kind and/or adding zero-inflation allows to take characteristics
of real data into account.
Additionally, one population and two population mixtures - with and
without zero-inflations - allow distribution fittings of the Poisson,
 negative binomial and the Poisson-beta distribution.
}
\details{
Functions nlogL_pois, nlogL_nb, nlogL_pb compute the negative
log-likelihood of poisson, negative binomial and the poisson-beta
distributions against the data.
Functions nlogL_pois2, nlogL_nb2 and nlogL_pb2 compute the negative
log-likelihood values for a bimodal mixture distribution whereas
nlogL_zipois, nlogL_zinb, nlogL_zipb compute the same for the
zero-inflated distributions. Furthermore, nlogL_zipois2, nlogL_zinb2
and nlogL_zipb2 are for bimodal distributions with zero-inflation.
}
\keyword{beta}
\keyword{binomial}
\keyword{likelihood}
\keyword{negative}
\keyword{poisson}
