% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_est_fns.R
\name{par-est-fns}
\alias{par-est-fns}
\alias{estimate_mpb_optim_init}
\alias{get_0inf_parameter}
\alias{get_fitted_params}
\title{Functions to compute parameter estimates}
\usage{
estimate_mpb_optim_init(x, iter = 200)

get_0inf_parameter(x)

get_fitted_params(x, type)
}
\arguments{
\item{x}{Gene expression data as an array}

\item{iter}{number of bootstrap replicates to estimate initial
parameters for the mpb}

\item{type}{keyword for the distribution the data is to be fitted
against. Possible values are ("pois", "zip", "nb", "zinb",
"mpb", "zimpb", "nb2", "mpb2")}
}
\description{
Functions to compute parameter estimates
}
\keyword{estimation}
\keyword{parameter}
